+++ moor_generator/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 1.2.0
 - Blob data type
 - Generated classes now use lazy getters instead of recalculating fields on each access
+- Custom Constraints
+- Data classes can be converted from and to json
 
 ## 1.1.0
 - The generated data classes now implement `toString()`